refactor: remove subdirectories for events
